+++ DeliveryDroneEcommerce/View Controllers/CompanyView/CompanyOrderViewController.swift
@@ -19,7 +19,7 @@ class CompanyOrderViewController: UIViewController {
       var time : String = ""
       var name : String = ""
       var address : String = ""
-      var price : Int = 0
+      var price : Double = 0
       var dist : Int = 0
     
     
@@ -65,7 +65,7 @@ class CompanyOrderViewController: UIViewController {
                 print("----------------")
                 print(companyID)
                 print("----------------")
-                let price = value["Price"] as! Int
+                let price = value["Price"] as! Double
                 let product = value["Product"] as! String
                 let place = value["Place"] as! Int
                 let desc = value["Description"] as! String
